+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Github templates for issues and PRs #29
 - Update box ~> 0.6.1 #31
 - Automate release process #26
+- Added switch for canary deployment #11
 
 ### Changed